test(setup): extract verified inference fixtures and narrow config types
